requirements/local.txt
-r ./base.txt

django-debug-toolbar==1.11  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.1.4  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.4.4  # https://github.com/pytest-dev/pytest-django
django-widget-tweaks==1.4.3
alabaster==0.7.12
amqp==2.4.1
argon2-cffi==18.3.0
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.3.2
defusedxml==0.5.0
django-allauth==0.38.0
django-appconf==1.0.2
django-compressor==2.2
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.4
django-model-utils==3.1.2
django-redis==4.10.0
django-session-security==2.6.5
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
dnspython==1.16.0
docutils==0.14
eventlet==0.24.1
factory-boy==2.11.1
Faker==1.0.2
flake8==3.6.0
greenlet==0.4.15
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
mpfshell==0.9.0
mypy==0.650
mypy-extensions==0.4.1
oauthlib==3.0.1
packaging==19.0
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.1
prompt-toolkit==2.0.8
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
pyparsing==2.3.1
pyserial==3.4
pytest==4.1.0
pytest-django==3.4.4
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.7
rcssmin==1.0.6
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
selenium==3.141.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.1.2
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
websocket-client==0.35.0
Werkzeug==0.14.1
django-dbbackup #https://django-dbbackup.readthedocs.io/en/stable/installation.html
#pip install -e git+https://github.com/mjs7231/django-dbbackup.git#egg=django-dbbackup
django-storages #
paramiko #
